fix player mongodb repository
