+++ utopia-gamification-quest/src/main/kotlin/tw/waterballsa/utopia/utopiagamificationquest/domain/quests/QuestBuilder.kt
@@ -16,16 +16,20 @@ class Quests(val wsa: WsaDiscordProperties) {
             listOf(
                 unlockAcademyQuest,
                 selfIntroductionQuest,
-                SendContainsImageMessageInEngineerLifeChannelQuest,
                 firstMessageActionQuest,
+                SendContainsImageMessageInEngineerLifeChannelQuest,
+                watchVideoQuest,
                 flagPostQuest,
                 ReplyToAnyoneInCareerAdvancementTopicChannelQuest,
+                SendMessageInVoiceChannelQuest,
+                JoinActivityQuest,
                 quizQuest
             )
         )
     }
 
     fun findById(questId: Int): Quest = quests.first { it.id == questId }
+
     fun String.toLink(): String = "https://discord.com/channels/${wsa.guildId}/${this}"
 
 }
