+++ utopia-gamification-quest/src/main/kotlin/tw/waterballsa/utopia/utopiagamificationquest/listeners/JoinActivityListener.kt
@@ -1,21 +1,18 @@
-package tw.waterballsa.utopia.utopiagamificationquest
+package tw.waterballsa.utopia.utopiagamificationquest.listeners
 
 import mu.KotlinLogging
+import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.ScheduledEvent.Status
-import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.entities.channel.unions.AudioChannelUnion
 import net.dv8tion.jda.api.events.guild.scheduledevent.ScheduledEventCreateEvent
 import net.dv8tion.jda.api.events.guild.scheduledevent.ScheduledEventDeleteEvent
 import net.dv8tion.jda.api.events.guild.scheduledevent.update.ScheduledEventUpdateStatusEvent
 import net.dv8tion.jda.api.events.guild.voice.GuildVoiceUpdateEvent
 import org.springframework.stereotype.Component
-import tw.waterballsa.utopia.jda.UtopiaListener
 import tw.waterballsa.utopia.utopiagamificationquest.domain.Activity
-import tw.waterballsa.utopia.utopiagamificationquest.domain.Audience
 import tw.waterballsa.utopia.utopiagamificationquest.domain.DateTimeRange
-import tw.waterballsa.utopia.utopiagamificationquest.domain.Mission
-import tw.waterballsa.utopia.utopiagamificationquest.extensions.claimMissionReward
 import tw.waterballsa.utopia.utopiagamificationquest.repositories.ActivityRepository
+import tw.waterballsa.utopia.utopiagamificationquest.repositories.PlayerRepository
 import tw.waterballsa.utopia.utopiagamificationquest.service.PlayerFulfillMissionsService
 import java.time.ZoneId
 
@@ -23,22 +20,25 @@ private val log = KotlinLogging.logger {}
 
 @Component
 class EventJoiningListener(
+    guild: Guild,
+    playerRepository: PlayerRepository,
     private val playerFulfillMissionsService: PlayerFulfillMissionsService,
     private val activityRepository: ActivityRepository
-) : UtopiaListener() {
+) : UtopiaGamificationListener(guild, playerRepository) {
 
     override fun onGuildVoiceUpdate(event: GuildVoiceUpdateEvent) {
         with(event) {
-            val player = member.user
+            val user = member.user
+            val player = user.toPlayer() ?: return
 
             channelJoined?.activity?.let {
-                it.join(player.toAudience())
+                it.join(player)
                 activityRepository.save(it)
             }
 
             channelLeft?.activity?.let {
-                val action = it.leave(player.id) ?: return
-                playerFulfillMissionsService.execute(action, player.presenter)
+                val action = it.leave(player) ?: return
+                playerFulfillMissionsService.execute(action, user.presenter)
                 activityRepository.save(it)
             }
         }
@@ -47,15 +47,6 @@ class EventJoiningListener(
     //TODO 如果 兩個活動 重複了 channel id 該怎麼辦
     private val AudioChannelUnion.activity
         get() = activityRepository.findInProgressActivityByChannelId(id)
-
-    private fun User.toAudience(): Audience = Audience(id, name)
-
-    private val User.presenter
-        get() = object : PlayerFulfillMissionsService.Presenter {
-            override fun presentClaimMissionReward(mission: Mission) {
-                claimMissionReward(mission)
-            }
-        }
 
     override fun onScheduledEventCreate(event: ScheduledEventCreateEvent): Unit =
         with(event.scheduledEvent) {
