+++ utopia-gamification-quest/src/main/kotlin/tw/waterballsa/utopia/utopiagamificationquest/listeners/PostListener.kt
@@ -1,5 +1,6 @@
-package tw.waterballsa.utopia.utopiagamificationquest
+package tw.waterballsa.utopia.utopiagamificationquest.listeners
 
+import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.events.guild.GuildAuditLogEntryCreateEvent
 import org.springframework.stereotype.Component
@@ -9,37 +10,33 @@ import tw.waterballsa.utopia.utopiagamificationquest.domain.Mission
 import tw.waterballsa.utopia.utopiagamificationquest.domain.Player
 import tw.waterballsa.utopia.utopiagamificationquest.domain.actions.PostAction
 import tw.waterballsa.utopia.utopiagamificationquest.extensions.claimMissionReward
+import tw.waterballsa.utopia.utopiagamificationquest.listeners.UtopiaGamificationListener
+import tw.waterballsa.utopia.utopiagamificationquest.repositories.PlayerRepository
 import tw.waterballsa.utopia.utopiagamificationquest.service.PlayerFulfillMissionsService
 
 @Component
 class PostListener(
+    guild: Guild,
+    playerRepository: PlayerRepository,
     private val properties: WsaDiscordProperties,
     private val playerFulfillMissionsService: PlayerFulfillMissionsService
-) : UtopiaListener() {
+) : UtopiaGamificationListener(guild, playerRepository) {
 
     override fun onGuildAuditLogEntryCreate(event: GuildAuditLogEntryCreateEvent) {
         with(event) {
-            val player = user ?: return
-            val action = action ?: return
+            val user = user ?: return
+            val channel = guild.getThreadChannelById(entry.targetId)?.parentChannel ?: return
+            val player = user.toPlayer() ?: return
 
-            playerFulfillMissionsService.execute(action, player.presenter)
+            val action = PostAction(
+                player,
+                channel.id
+            )
+
+            playerFulfillMissionsService.execute(action, user.presenter)
         }
     }
 
     private val GuildAuditLogEntryCreateEvent.user get() = entry.user ?: jda.retrieveUserById(entry.userId).complete()
 
-    private val GuildAuditLogEntryCreateEvent.action
-        get() = guild.getThreadChannelById(entry.targetId)?.let {
-            PostAction(
-                Player(user.id, user.name),
-                it.parentChannel.id
-            )
-        }
-
-    private val User.presenter
-        get() = object : PlayerFulfillMissionsService.Presenter {
-            override fun presentClaimMissionReward(mission: Mission) {
-                claimMissionReward(mission)
-            }
-        }
 }
