+++ wsa-bot-commands.md
@@ -5,11 +5,6 @@
 |:--------:| --------- | --------------------------- |
 |   1a2b   |           | Start a new guess num game. |
 
-## rock-paper-scissors
-|      Commands       | Arguments | Description                           |
-|:-------------------:| --------- | ------------------------------------- |
-| rock-paper-scissors |           | start a new rock paper scissors game! |
-
 ## gaas
 |        Commands        | Arguments                                                                                          | Description                                          |
 |:----------------------:| -------------------------------------------------------------------------------------------------- | ---------------------------------------------------- |
@@ -43,15 +38,15 @@
 |:--------------:| ---------------------------------------------------------------------------------------------------------- | ----------- |
 | random lottery | number(INTEGER): Number of choose members per room.<br>role(ROLE): Only select specific role in this round | Lottery     |
 
-## roulette
-| Commands | Arguments | Description     |
-|:--------:| --------- | --------------- |
-| roulette |           | Start the game. |
-
 ## utopia
 | Commands | Arguments | Description    |
 |:--------:| --------- | -------------- |
 |  utopia  |           | utopia command |
+
+## roulette
+| Commands | Arguments | Description     |
+|:--------:| --------- | --------------- |
+| roulette |           | Start the game. |
 
 ## message
 |        Commands        | Arguments                                                                                                               | Description                               |
